+++ CHANGES.md
@@ -6,6 +6,7 @@ that have the Hip2 file-write error. Note that this is not perfect, there are so
 - Note that this write-out bug fix does not work on the dvd data (note, 
   the dvd still has the same bug, it is just not easily correctable).
 - Updated the flagged source list (hip2_javatool_flagged.txt).
+- Parallax factors from the IAD are now loaded on data.parse()
 
 0.3.5 (2021-04-19)
 ------------------